test: split LatticePageTest into per-domain endpoint, serialization and page tests
